(refactor) Update import statements with new file locations.
